Ocvdnn component (#123)
* Use generalized ReadNet
* Add TensorFlow vehicle color model
* Remove support for Rachmadi's vehicle color model
* Add capability to specify model input name
