+++ cpp/CMakeLists.txt
@@ -33,7 +33,7 @@ set(CMAKE_CXX_STANDARD 11)
 add_subdirectory(OcvFaceDetection)
 add_subdirectory(DlibFaceDetection)
 add_subdirectory(OalprLicensePlateTextDetection)
-add_subdirectory(CaffeDetection)
+add_subdirectory(OcvDnnDetection)
 add_subdirectory(DarknetDetection)
 add_subdirectory(TesseractOCRTextDetection)
 add_subdirectory(SceneChangeDetection)
