--- cpp/CaffeDetection/plugin-files/models/models.ini
@@ -1,21 +1,0 @@
-[googlenet]
-# The following 3 files are provided in the component source code.
-# This model can also be found at https://github.com/BVLC/caffe/tree/master/models
-model_txt=bvlc_googlenet.prototxt
-model_bin=bvlc_googlenet.caffemodel
-synset_txt=synset_words.txt
-
-[yahoo_nsfw]
-# The following 3 files are provided in the component source code.
-# This model can also be found at https://github.com/yahoo/open_nsfw
-model_txt=yahoo_nsfw.prototxt
-model_bin=yahoo_nsfw.caffemodel
-synset_txt=yahoo_nsfw.txt
-
-[vehicle_color]
-# The synset_txt file is provided in the component source code.
-# The model prototxt can be found at https://github.com/rezafuad/vehicle-color-recognition/blob/master/vcolor_deploy02.prototxt
-# Instructions for downloading the caffemodel can be found at https://github.com/rezafuad/vehicle-color-recognition
-model_txt=vehicle_color.prototxt
-model_bin=vehicle_color.caffemodel
-synset_txt=vehicle_color.txt
