--- cpp/CaffeDetection/plugin-files/models/vehicle_color.txt
@@ -1,8 +1,0 @@
-0 yellow
-1 white
-2 blue
-3 cyan
-4 red
-5 gray
-6 black
-7 green
