--- cpp/CaffeDetection/plugin-files/models/yahoo_nsfw.txt
@@ -1,2 +1,0 @@
-0 safe for work
-1 not safe for work
