--- cpp/CaffeDetection/test/data/NOTICE
@@ -1,15 +1,0 @@
-# digital-clock.jpg
-Public Domain
-https://en.wikipedia.org/wiki/Clock#/media/File:Digital-clock-radio-basic_hf.jpg
-
-# sundial.jpg
-Public Domain
-https://en.wikipedia.org/wiki/Sundial#/media/File:Sundial_in_Supreme_Court_Gardens,_Perth.jpg
-
-# ff-region-object-motion.avi
-Image on left side of frame is digital-clock.jpg
-Public Domain
-Image on right side of frame is sundial.jpg
-Public Domain
-No motion in first 30 frames.
-Starting at frame 31 image on left moves.
