+++ cpp/OcvDnnDetection/NOTICE
@@ -1,0 +1,7 @@
+# NOTICE
+
+This software (or technical data) was produced for the U.S. Government
+under contract, and is subject to the Rights in Data-General Clause
+52.227-14, Alt. IV (DEC 2007).
+
+Copyright 2019 The MITRE Corporation. All Rights Reserved.
