+++ cpp/OcvDnnDetection/plugin-files/config/NOTICE
@@ -1,0 +1,21 @@
+# bvlc_googlenet.caffemodel
+# bvlc_googlenet.prototxt
+Unrestricted use.
+https://github.com/BVLC/caffe/tree/master/models/bvlc_googlenet
+
+
+# synset_words.txt
+Unrestricted use.
+http://dl.caffe.berkeleyvision.org/caffe_ilsvrc12.tar.gz
+
+
+# yahoo_nsfw.caffemodel
+Copyright 2016, Yahoo Inc.
+https://github.com/yahoo/open_nsfw
+Renamed from "resnet_50_1by2_nsfw.caffemodel".
+
+
+# yahoo_nsfw.prototxt
+Copyright 2016, Yahoo Inc.
+https://github.com/yahoo/open_nsfw
+Renamed from "deploy.prototxt".
