+++ cpp/OcvDnnDetection/plugin-files/models/models.ini
@@ -1,0 +1,19 @@
+[googlenet]
+# The following 3 files are provided in the component source code.
+# This model can also be found at https://github.com/BVLC/caffe/tree/master/models
+model_config=bvlc_googlenet.prototxt
+model_binary=bvlc_googlenet.caffemodel
+synset_txt=synset_words.txt
+
+[yahoo_nsfw]
+# The following 3 files are provided in the component source code.
+# This model can also be found at https://github.com/yahoo/open_nsfw
+model_config=yahoo_nsfw.prototxt
+model_binary=yahoo_nsfw.caffemodel
+synset_txt=yahoo_nsfw.txt
+
+[vehicle_color]
+# The following 2 files are provided in the component source code.
+# vehicle_color.pbtxt is also included for reference.
+model_binary=vehicle_color.pb
+synset_txt=vehicle_color.txt
