+++ cpp/OcvDnnDetection/plugin-files/models/vehicle_color.txt
@@ -1,0 +1,11 @@
+0 black
+1 blue
+2 brown
+3 gray
+4 green
+5 orange
+6 pink
+7 purple
+8 red
+9 white
+10 yellow
