+++ cpp/OcvDnnDetection/test/data/NOTICE
@@ -1,0 +1,27 @@
+# digital-clock.jpg
+Public Domain
+https://en.wikipedia.org/wiki/Clock#/media/File:Digital-clock-radio-basic_hf.jpg
+
+# sundial.jpg
+Public Domain
+https://en.wikipedia.org/wiki/Sundial#/media/File:Sundial_in_Supreme_Court_Gardens,_Perth.jpg
+
+# ff-region-object-motion.avi
+Image on left side of frame is digital-clock.jpg
+Public Domain
+Image on right side of frame is sundial.jpg
+Public Domain
+No motion in first 30 frames.
+Starting at frame 31 image on left moves.
+
+yellow-car.jpg
+Public Domain
+https://free-images.com/display/taxi_car_on_new.html
+
+red-car.jpg
+Public Domain
+https://free-images.com/display/car_red_car_vehicle.html
+
+blue-car.jpg
+Public Domain
+https://commons.wikimedia.org/wiki/File:Car_civic.jpg
